created module and cleaned up a little
